Added new functional: - ability to purchase item - added cart of goods

+++ LibraryApp.Presentation/Presenters/MainPresenter.cs
@@ -3,6 +3,7 @@ using LibraryApp.Core.Entities;
 using LibraryApp.Presentation.ViewContracts;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,12 +13,14 @@ namespace LibraryApp.Presentation.Presenters
     public class MainPresenter : BasePresenter<IMainView>
     {
         private readonly IRepository _repo;
+        private List<IProduct> _list;
 
         public MainPresenter(IApplicationController controller, IMainView view, IRepository repository)
             :base(controller, view)
         {
             _repo = repository;
             LoadData();
+            View.Cart += OnClick_Cart;
 
             View.AddBook += OnClick_AddBook;
             View.AddMagazine += OnClick_AddMagazine;
@@ -30,44 +33,68 @@ namespace LibraryApp.Presentation.Presenters
             View.EditBook += OnClick_EditBook;
             View.EditMagazine += OnClick_EditMagz;
             View.EditNewspaper += OnClick_EditNewsp;
+
+            View.AddBookToCart += () => OnCellClick_AddBookToCart(View.BookId);
+            View.AddMagazineToCart += () => OnCellClick_AddMagazineToCart(View.MagazId);
+            View.AddNewspaperToCart += () => OnCellClick_AddNewspaperToCart(View.NewspaperId);
+
+            _list = new List<IProduct>();
         }
 
         private void LoadData()
         {
             View.Load(_repo.Books, _repo.Magazines, _repo.Newspapers);
         }
+        private void OnClick_Cart()
+        {
+            Controller.Run<CartPresenter, List<IProduct>>(_list);
+        }
 
         private void OnClick_AddBook()
         {
-            Controller.Run<AddBookPresenter>();
+            Controller.Run<AddBookPresenter, IRepository>(_repo);
             View.Message("The book is added successfully!");
         }
-
         private void OnClick_AddMagazine()
         {
-            Controller.Run<AddMagazinePresenter>();
+            Controller.Run<AddMagazinePresenter, IRepository>(_repo);
             View.Message("The magazine is added successfully!");
         }
         private void OnClick_AddNewspaper()
         {
-            Controller.Run<AddNewspPresenter>();
+            Controller.Run<AddNewspPresenter, IRepository>(_repo);
             View.Message("The newspaper is added successfully!");
         }
-
+   
         private void OnClick_DeleteBook(int id)
         {
+            if (id == 0)
+            {
+                View.Message("The list is empty.");
+                return;
+            }
             var book = _repo.Books.First(b => b.ID == id);
             _repo.Delete(book);
             View.Message("It's deleted.");
         }
         private void OnClick_DeleteMagazine(int id)
         {
+            if (id == 0)
+            {
+                View.Message("The list is empty.");
+                return;
+            }
             var magazine = _repo.Magazines.First(m => m.ID == id);
             _repo.Delete(magazine);
             View.Message("It's deleted.");
         }
         private void OnClick_DeleteNewspaper(int id)
         {
+            if (id == 0)
+            {
+                View.Message("The list is empty.");
+                return;
+            }
             var paper = _repo.Newspapers.First(n => n.ID == id);
             _repo.Delete(paper);
             View.Message("It's deleted.");
@@ -89,5 +116,38 @@ namespace LibraryApp.Presentation.Presenters
             View.Message("Saved successfully.");
         }
 
+        private void OnCellClick_AddBookToCart(int id)
+        {
+            var book = _repo.Books.First(b => b.ID == id);
+            if (_list.Contains(book))
+            {
+                book.ProductCount += 1;
+                return;
+            }
+            book.ProductCount = 1;
+            _list.Add(book);
+        }
+        private void OnCellClick_AddMagazineToCart(int id)
+        {
+            var mag = _repo.Magazines.First(m => m.ID == id);
+            if (_list.Contains(mag))
+            {
+                mag.ProductCount += 1;
+                return;
+            }
+            mag.ProductCount = 1;
+            _list.Add(mag);
+        }
+        private void OnCellClick_AddNewspaperToCart(int id)
+        {
+            var newsp = _repo.Newspapers.First(n => n.ID == id);
+            if (_list.Contains(newsp))
+            {
+                newsp.ProductCount += 1;
+                return;
+            }
+            newsp.ProductCount = 1;
+            _list.Add(newsp);
+        }
     }
 }
